# Increase neighbors search radius for corners in ChessBoardDetector:findQuadNeighbors